+++ frontend/src/views/MyProgress/information/CanteenInfo.vue
