+++ frontend/src/views/MyProgress/information/DiversificationMeasureInfo.vue
