+++ frontend/src/views/MyProgress/information/WasteMeasureInfo.vue
